requirements.txt
torch==1.8.1
torchvision==0.9.1
Pillow==8.1.2
numpy==1.20.2
nltk==3.6.2
pycocotools==2.0.2
streamlit==0.81.1
